rag/requirements.txt
streamlit==1.33.0

langchain==0.1.16

python-dotenv==1.0.1
requests==2.31.0

sentence-transformers==2.6.1
torch>=2.0.0

faiss-cpu==1.7.4
pypdf==4.1.0

numpy>=1.24.0
tqdm==4.66.2
